refactor with remarks from reviewer Kkh
