+++ mentionserver/src/main/java/com.mention/controller/UserController.java
@@ -1,46 +1,31 @@
 package com.mention.controller;
 
-import com.mention.model.User;
-import com.mention.service002.UserService;
+import com.mention.dto.UserDtoIdRq;
+import com.mention.dto.UserDtoRq;
+import com.mention.service.UserServiceImpl;
 import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-
-import java.util.Optional;
 
 @RestController
 @RequestMapping("/api/user")
 public class UserController {
 
-  private UserService userService;
+  private UserServiceImpl userService;
 
-  public UserController(UserService userService) {
+  public UserController(UserServiceImpl userService) {
     this.userService = userService;
   }
 
-  @PostMapping
-  public void addUser(@RequestBody User user) {
-    userService.addUser(user);
+  @PostMapping("/add")
+  public void createUser(@RequestBody UserDtoRq userDtoNewUser) {
+    userService.createNewUser(userDtoNewUser);
   }
 
-  @GetMapping(value = "/{id}")
-  public Optional<User> getUser(@PathVariable Long id) {
-    return userService.getUser(id);
+  @DeleteMapping("/delete")
+  public  void deleteUser(@RequestBody UserDtoIdRq userDtoIdRq) {
+    userService.deleteUser(userDtoIdRq);
   }
-
-  @PutMapping
-  public void updateUser(@RequestBody User user) {
-    userService.updateUser(user);
-  }
-
-  @DeleteMapping(value = "/{id}")
-  public void deleteUser(@PathVariable Long id) {
-    userService.deleteUser(id);
-  }
-
 }
