+++ mentionserver/src/main/java/com.mention/dto/UserDtoRq.java
@@ -3,13 +3,14 @@ package com.mention.dto;
 import lombok.Data;
 
 @Data
-public class UserDtoNewUserRq {
+public class UserDtoRq {
 
   private String username;
-  private String email;
-  private String password;
-  // we need parameter 'isActive' in this Dto ?
-  private boolean isActive;
 
+  private String email;
+
+  private String password;
+
+  private boolean isActive;
 
 }
