+++ mentionserver/src/main/java/com.mention/service/UserService.java
@@ -1,0 +1,12 @@
+package com.mention.service;
+
+import com.mention.dto.UserDtoIdRq;
+import com.mention.dto.UserDtoRq;
+
+public interface UserService {
+
+  void createNewUser(UserDtoRq userDtoNewUser);
+
+  void deleteUser(UserDtoIdRq user);
+
+  }
