+++ mentionserver/src/main/java/com.mention/service/UserServiceImpl.java
@@ -1,9 +1,9 @@
 package com.mention.service;
 
 import com.mention.dto.UserDtoIdRq;
-import com.mention.dto.UserDtoNewUserRq;
-import com.mention.model.User;
+import com.mention.dto.UserDtoRq;
 import com.mention.repository.UserRepository;
+import com.mention.model.User;
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -12,21 +12,21 @@ import org.springframework.transaction.annotation.Transactional;
 import java.util.Optional;
 
 @Service
-public class UserNewUserImpl implements UserNewUser {
+public class UserServiceImpl implements UserService {
+
 
   private UserRepository userRepository;
 
   @Autowired
-  public UserNewUserImpl(UserRepository userRepository) {
+  public UserServiceImpl(UserRepository userRepository) {
     this.userRepository = userRepository;
   }
 
   @Override
   @Transactional
-  public void createNewUser(UserDtoNewUserRq userDtoNewUser) {
+  public void createNewUser(UserDtoRq userDtoNewUser) {
     ModelMapper modelMapper = new ModelMapper();
     User insertUser = modelMapper.map(userDtoNewUser, User.class);
-    insertUser.setActive(true);
     userRepository.save(insertUser);
   }
 
@@ -35,7 +35,7 @@ public class UserNewUserImpl implements UserNewUser {
   public void deleteUser(UserDtoIdRq user) {
     ModelMapper modelMapper = new ModelMapper();
     Optional<User> currentUser = userRepository.findById(user.getId());
-    if (currentUser.isPresent() && currentUser.get().isActive()) {
+    if (currentUser.isPresent()) {
       currentUser.get().setActive(false);
       userRepository.save(currentUser.get());
     }
