--- mentionserver/src/main/java/com.mention/service002/UserService.java
@@ -1,16 +1,0 @@
-package com.mention.service002;
-
-import com.mention.model.User;
-
-import java.util.Optional;
-
-public interface UserService {
-
-  void addUser(User user);
-
-  Optional<User> getUser(Long id);
-
-  void updateUser(User user);
-
-  void deleteUser(Long id);
-}
